@@ -66,7 +66,10 @@
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <jackson.version>2.5.0</jackson.version>
         <guava.version>18.0</guava.version>
-        <slf4j.version>1.7.10</slf4j.version>
+        <maven.compiler.source>13</maven.compiler.source>
+        <maven.compiler.target>13</maven.compiler.target>
+        <maven.javadoc.source>13</maven.javadoc.source>
+        <maven.surefire.source>3.0.0-M4</maven.surefire.source>
     </properties>
     <dependencies>
         <dependency>
@@ -181,10 +190,11 @@
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-compiler-plugin</artifactId>
-                <version>2.3.2</version>
+                <version>3.8.1</version>
                 <configuration>
-                    <source>1.6</source>
-                    <target>1.6</target>
+                    <source>${maven.compiler.source}</source>
+                    <target>${maven.compiler.target}</target>
+                    <release>13</release>
                 </configuration>
             </plugin>
             <plugin>
@@ -294,19 +304,19 @@
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-javadoc-plugin</artifactId>
-                <version>2.8</version>
+                <version>3.1.1</version>
                 <configuration>
-                    <additionalparam>-quiet -Xdoclint:none</additionalparam>
                     <!--
                       | Apple's JVM sometimes requires more memory
                     -->
                     <additionalJOption>-J-Xmx1024m</additionalJOption>
-                    <source>1.6</source>
+                    <source>13</source>
                     <encoding>UTF-8</encoding>
                     <maxmemory>2g</maxmemory>
                     <links>
-                        <link>http://commons.apache.org/lang/api/</link>
-                        <link>http://download.oracle.com/javase/6/docs/api/</link>
+                        <link>http://commons.apache.org/dbcp/apidocs/</link>
+                        <link>http://commons.apache.org/fileupload/apidocs/</link>
+                        <link>https://docs.oracle.com/en/java/javase/13/docs/api/</link>
                     </links>
                     <linksource>true</linksource>
                     <show>public</show>
