@@ -151,6 +151,9 @@ public class RecurlyClient {
 
     /**
      * Open the underlying http client
+     *
+     * @throws NoSuchAlgorithmException throws NoSuchAlgorithmException
+     * @throws KeyManagementException throws KeyManagementException
      */
     public synchronized void open() throws NoSuchAlgorithmException, KeyManagementException {
         client = createHttpClient();
@@ -432,6 +435,7 @@ public class RecurlyClient {
      * postpone a subscription, setting a new renewal date.
      *
      * @param subscription Subscription object
+     * @param renewaldate DateTime renewalDate
      * @return -?-
      */
     public Subscription postponeSubscription(final Subscription subscription, final DateTime renewaldate) {
@@ -443,6 +447,7 @@ public class RecurlyClient {
      * Terminate a particular {@link Subscription} by it's UUID
      *
      * @param subscription Subscription to terminate
+     * @param refund RefundOption object
      */
     public void terminateSubscription(final Subscription subscription, final RefundOption refund) {
         doPUT(Subscription.SUBSCRIPTION_RESOURCE + "/" + subscription.getUuid() + "/terminate?refund=" + refund,
@@ -484,6 +489,7 @@ public class RecurlyClient {
      * Returns information about a single subscription.
      *
      * @param uuid UUID of the subscription to preview an update for
+     * @param subscriptionUpdate SubscriptionUpdate object
      * @return Subscription the updated subscription preview
      */
     public Subscription updateSubscriptionPreview(final String uuid, final SubscriptionUpdate subscriptionUpdate) {
@@ -830,7 +836,6 @@ public class RecurlyClient {
 
     /**
      * Lookup an account's invoices
-     * <p>
      * Returns the account's invoices
      *
      * @param accountCode recurly account id
@@ -843,7 +848,6 @@ public class RecurlyClient {
 
     /**
      * Refund an invoice given an open amount
-     * <p/>
      * Returns the refunded invoice
      *
      * @param invoiceId The id of the invoice to refund
@@ -861,7 +865,6 @@ public class RecurlyClient {
 
     /**
      * Refund an invoice given some line items
-     * <p/>
      * Returns the refunded invoice
      *
      * @param invoiceId The id of the invoice to refund
@@ -879,7 +882,6 @@ public class RecurlyClient {
 
     /**
      * Lookup an account's shipping addresses
-     * <p>
      * Returns the account's shipping addresses
      *
      * @param accountCode recurly account id
@@ -892,7 +894,6 @@ public class RecurlyClient {
 
     /**
      * Lookup an account's invoices given query params
-     * <p>
      * Returns the account's invoices
      *
      * @param accountCode recurly account id
@@ -911,7 +912,8 @@ public class RecurlyClient {
      * <p>
      * Returns an invoice
      *
-     * @param accountCode
+     * @param accountCode the accountCode
+     * @param invoice the Invoice
      * @return the invoice that was generated on success, null otherwise
      */
     public Invoice postAccountInvoice(final String accountCode, final Invoice invoice) {
@@ -924,6 +926,7 @@ public class RecurlyClient {
      * @deprecated Prefer using Invoice#getId() as the id param (which is a String)
      *
      * @param invoiceId Recurly Invoice ID
+     * @return invoice Invoice Object
      */
     @Deprecated
     public Invoice markInvoiceSuccessful(final Integer invoiceId) {
@@ -934,6 +937,7 @@ public class RecurlyClient {
      * Mark an invoice as paid successfully - Recurly Enterprise Feature
      *
      * @param invoiceId String Recurly Invoice ID
+     * @return invoice Invoice Object
      */
     public Invoice markInvoiceSuccessful(final String invoiceId) {
         return doPUT(Invoices.INVOICES_RESOURCE + "/" + invoiceId + "/mark_successful", null, Invoice.class);
@@ -945,6 +949,7 @@ public class RecurlyClient {
      * @deprecated Prefer using Invoice#getId() as the id param (which is a String)
      *
      * @param invoiceId Recurly Invoice ID
+     * @return invoice Invoice Object
      */
     @Deprecated
     public Invoice markInvoiceFailed(final Integer invoiceId) {
@@ -955,6 +960,7 @@ public class RecurlyClient {
      * Mark an invoice as failed collection
      *
      * @param invoiceId String Recurly Invoice ID
+     * @return invoice Invoice Object
      */
     public Invoice markInvoiceFailed(final String invoiceId) {
         return doPUT(Invoices.INVOICES_RESOURCE + "/" + invoiceId + "/mark_failed", null, Invoice.class);
@@ -964,6 +970,7 @@ public class RecurlyClient {
      * Force collect an invoice
      *
      * @param invoiceId String Recurly Invoice ID
+     * @return invoice Invoice Object
      */
     public Invoice forceCollectInvoice(final String invoiceId) {
         return doPUT(Invoices.INVOICES_RESOURCE + "/" + invoiceId + "/collect", null, Invoice.class);
@@ -976,6 +983,7 @@ public class RecurlyClient {
      *
      * @param invoiceId Recurly Invoice ID
      * @param payment   The external payment
+     * @return Transaction transaction Object
      */
     @Deprecated
     public Transaction enterOfflinePayment(final Integer invoiceId, final Transaction payment) {
@@ -987,6 +995,7 @@ public class RecurlyClient {
      *
      * @param invoiceId String Recurly Invoice ID
      * @param payment   The external payment
+     * @return transaction Transaction Object
      */
     public Transaction enterOfflinePayment(final String invoiceId, final Transaction payment) {
         return doPOST(Invoices.INVOICES_RESOURCE + "/" + invoiceId + "/transactions", payment, Transaction.class);
@@ -1110,7 +1119,7 @@ public class RecurlyClient {
      * Return all the {@link AddOn} for a {@link Plan}
      * <p>
      *
-     * @param planCode
+     * @param planCode the plan code
      * @return the {@link AddOn} objects as identified by the passed plan ID
      */
     public AddOns getAddOns(final String planCode) {
@@ -1125,7 +1134,7 @@ public class RecurlyClient {
      * Return all the {@link AddOn} for a {@link Plan}
      * <p>
      *
-     * @param planCode
+     * @param planCode the plan code
      * @param params {@link QueryParams}
      * @return the {@link AddOn} objects as identified by the passed plan ID
      */
@@ -1194,6 +1203,7 @@ public class RecurlyClient {
      * Redeem a {@link Coupon} on an account.
      *
      * @param couponCode redeemed coupon id
+     * @param redemption Redemption Object
      * @return the {@link Coupon} object
      */
     public Redemption redeemCoupon(final String couponCode, final Redemption redemption) {
